server/ai-service/pyproject.toml
[tool.poetry]
name = "ai-service"
version = "1.0.0"
description = "Quiz Generator and Summarizer Module using OCR and NLP."
authors = ["Muhammad Bilal Khan <@Hi-kue>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.13.2"
flask-cors = "^5.0.1"
flask-limiter = "^3.12"
flask = "^3.1.0"
rich = "^13.9.4"
pydantic = "^2.10.6"
python-dotenv = "^1.1.0"
requests = "^2.32.3"
marshmallow = "^3.26.1"
tiktoken = "^0.9.0"
docling = "^2.28.1"
google-genai = "^1.7.0"
flask-smorest = "^0.45.0"
fitz = "^0.0.1.dev2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
